requirements.txt
chromadb==0.5.11
google-ai-generativelanguage==0.6.6
google-generativeai==0.7.2
huggingface-hub==0.24.5
langchain==0.2.12
langchain-community==0.2.11
langchain-google-genai==1.0.8
langchainhub==0.1.20
python-dotenv==1.0.1
sentence-transformers==3.1.1
tiktoken==0.7.0
transformers==4.44.0
unstructured==0.15.3